--- src/main/java/com/weShare/api/v1/auth/login/CodeNotFoundException.java
@@ -1,7 +1,0 @@
-package com.weShare.api.v1.auth.login;
-
-public class CodeNotFoundException extends RuntimeException{
-    public CodeNotFoundException(String message) {
-        super(message);
-    }
-}
